requirements.txt
python-telegram-bot==13.15
gspread==5.9.0